requirements.txt
python-chess>=1.999
pandas>=2.0